requirements.txt
opencv-python>=4.5.3
python-chess>=1.9.4
watchdog>=2.1.0
numpy>=1.19.0

